+++ incapsula/client.go
@@ -36,7 +36,7 @@ func (c *Client) Verify() (*AccountStatusResponse, error) {
 	reqURL := fmt.Sprintf("%s/%s", c.config.BaseURL, endpointAccountStatus)
 	data := url.Values{}
 
-	resp, err := c.PostFormWithHeaders(reqURL, data)
+	resp, err := c.PostFormWithHeaders(reqURL, data, VerifyAccount)
 	if err != nil {
 		return nil, fmt.Errorf("Error checking account: %s", err)
 	}
@@ -67,36 +67,35 @@ func (c *Client) Verify() (*AccountStatusResponse, error) {
 	if resString != "0" {
 		return &accountStatusResponse, fmt.Errorf("Error from Incapsula service when checking account: %s", string(responseBody))
 	}
-
 	return &accountStatusResponse, nil
 }
 
-func (c *Client) PostFormWithHeaders(url string, data url.Values) (*http.Response, error) {
+func (c *Client) PostFormWithHeaders(url string, data url.Values, operation string) (*http.Response, error) {
 	req, err := http.NewRequest(http.MethodPost, url, strings.NewReader(data.Encode()))
 	if err != nil {
 		return nil, fmt.Errorf("Error preparing request: %s", err)
 	}
 
-	SetHeaders(c, req, contentTypeApplicationUrlEncoded, nil)
+	SetHeaders(c, req, contentTypeApplicationUrlEncoded, operation, nil)
 	return c.httpClient.Do(req)
 }
 
-func (c *Client) DoJsonRequestWithCustomHeaders(method string, url string, data []byte, headers map[string]string) (*http.Response, error) {
+func (c *Client) DoJsonRequestWithCustomHeaders(method string, url string, data []byte, headers map[string]string, operation string) (*http.Response, error) {
 	req, err := PrepareJsonRequest(method, url, data)
 	if err != nil {
 		return nil, fmt.Errorf("Error preparing request: %s", err)
 	}
 
-	SetHeaders(c, req, contentTypeApplicationJson, headers)
+	SetHeaders(c, req, contentTypeApplicationJson, operation, headers)
 
 	return c.httpClient.Do(req)
 }
 
-func (c *Client) DoJsonRequestWithHeaders(method string, url string, data []byte) (*http.Response, error) {
-	return c.DoJsonRequestWithCustomHeaders(method, url, data, nil)
+func (c *Client) DoJsonRequestWithHeaders(method string, url string, data []byte, operation string) (*http.Response, error) {
+	return c.DoJsonRequestWithCustomHeaders(method, url, data, nil, operation)
 }
 
-func (c *Client) DoJsonAndQueryParamsRequestWithHeaders(method string, url string, data []byte, params map[string]string) (*http.Response, error) {
+func (c *Client) DoJsonAndQueryParamsRequestWithHeaders(method string, url string, data []byte, params map[string]string, operation string) (*http.Response, error) {
 	req, err := PrepareJsonRequest(method, url, data)
 	if err != nil {
 		return nil, fmt.Errorf("Error preparing request: %s", err)
@@ -107,7 +106,7 @@ func (c *Client) DoJsonAndQueryParamsRequestWithHeaders(method string, url strin
 	}
 	req.URL.RawQuery = q.Encode()
 
-	SetHeaders(c, req, contentTypeApplicationJson, nil)
+	SetHeaders(c, req, contentTypeApplicationJson, operation, nil)
 
 	return c.httpClient.Do(req)
 }
@@ -123,13 +122,13 @@ func GetRequestParamsWithCaid(accountId int) map[string]string {
 	return params
 }
 
-func (c *Client) DoJsonRequestWithHeadersForm(method string, url string, data []byte, contentType string) (*http.Response, error) {
+func (c *Client) DoJsonRequestWithHeadersForm(method string, url string, data []byte, contentType string, operation string) (*http.Response, error) {
 	req, err := PrepareJsonRequest(method, url, data)
 	if err != nil {
 		return nil, fmt.Errorf("Error preparing request: %s", err)
 	}
 
-	SetHeaders(c, req, contentType, nil)
+	SetHeaders(c, req, contentType, operation, nil)
 	return c.httpClient.Do(req)
 }
 
@@ -141,11 +140,12 @@ func PrepareJsonRequest(method string, url string, data []byte) (*http.Request, 
 	return http.NewRequest(method, url, bytes.NewReader(data))
 }
 
-func SetHeaders(c *Client, req *http.Request, contentType string, customHeaders map[string]string) {
+func SetHeaders(c *Client, req *http.Request, contentType string, operation string, customHeaders map[string]string) {
 	req.Header.Set("Content-Type", contentType)
 	req.Header.Set("x-api-id", c.config.APIID)
 	req.Header.Set("x-api-key", c.config.APIKey)
 	req.Header.Set("x-tf-provider-ver", c.providerVersion)
+	req.Header.Set("x-tf-operation", operation)
 
 	if customHeaders != nil {
 		for name, value := range customHeaders {