+++ incapsula/client_notification_center_policy.go
@@ -65,7 +65,7 @@ func (c *Client) AddNotificationCenterPolicy(notificationPolicyFullDto *Notifica
 	}
 
 	log.Printf("[DEBUG] Add NotificationCenterPolicy with params %s and JSON request: %s\n", params, string(policyJSON))
-	resp, err := c.DoJsonAndQueryParamsRequestWithHeaders(http.MethodPost, reqURL, policyJSON, params)
+	resp, err := c.DoJsonAndQueryParamsRequestWithHeaders(http.MethodPost, reqURL, policyJSON, params, CreateNotificationCenterPolicy)
 	log.Printf("[DEBUG] client_notification_center_policy Post rest response:\n%+v", resp)
 	if err != nil {
 		return nil, fmt.Errorf("Error from NotificationCenter service when adding policy: %s ", err)
@@ -102,7 +102,7 @@ func (c *Client) UpdateNotificationCenterPolicy(notificationPolicyFullDto *Notif
 	params := GetRequestParamsWithCaid(notificationPolicyFullDto.AccountId)
 
 	log.Printf("[DEBUG] Update NotificationCenterPolicy JSON request: %s\n", string(policyJSON))
-	resp, err := c.DoJsonAndQueryParamsRequestWithHeaders(http.MethodPut, reqURL, policyJSON, params)
+	resp, err := c.DoJsonAndQueryParamsRequestWithHeaders(http.MethodPut, reqURL, policyJSON, params, UpdateNotificationCenterPolicy)
 	log.Printf("[DEBUG] client_notification_center_policy Put rest response:\n%+v", resp)
 	if err != nil {
 		return nil, fmt.Errorf("Error from NotificationCenter service when updateing policy: %s ", err)
@@ -129,7 +129,7 @@ func (c *Client) DeleteNotificationCenterPolicy(policyId int, accountId int) err
 	log.Printf("[INFO] Deleting NotificationCenterPolicy with ID %d ", policyId)
 	requestUrl := getRequestUrlWithId(c, policyId)
 	params := GetRequestParamsWithCaid(accountId)
-	resp, err := c.DoJsonAndQueryParamsRequestWithHeaders(http.MethodDelete, requestUrl, nil, params)
+	resp, err := c.DoJsonAndQueryParamsRequestWithHeaders(http.MethodDelete, requestUrl, nil, params, DeleteNotificationCenterPolicy)
 	log.Printf("[DEBUG] client_notification_center_policy Delete rest response:\n%+v", resp)
 	if err != nil {
 		return fmt.Errorf("Error from NotificationCenterPolicy service when deleting Policy with Id %d: %s ", policyId, err)
@@ -162,7 +162,7 @@ func (c *Client) GetNotificationCenterPolicy(policyId int, accountId int) (*Noti
 	requestUrl := getRequestUrlWithId(c, policyId)
 
 	params := GetRequestParamsWithCaid(accountId)
-	resp, err := c.DoJsonAndQueryParamsRequestWithHeaders(http.MethodGet, requestUrl, nil, params)
+	resp, err := c.DoJsonAndQueryParamsRequestWithHeaders(http.MethodGet, requestUrl, nil, params, ReadNotificationCenterPolicy)
 	log.Printf("[DEBUG] client_notification_center_policy Get rest response:\n%+v", resp)
 	if err != nil {
 		return nil, fmt.Errorf("Error from NotificationCenter service when reading policy with Id %d: %s ", policyId, err)