+++ incapsula/operation_constants.go
@@ -1,0 +1,105 @@
+package incapsula
+
+const VerifyAccount = "verify_account"
+
+const CreateSite = "create_site"
+const ReadSite = "read_site"
+const UpdateSite = "update_site"
+const DeleteSite = "delete_site"
+
+const CreatePolicy = "create_policy"
+const ReadPolicy = "read_policy"
+const UpdatePolicy = "update_policy"
+const DeletePolicy = "delete_policy"
+
+const CreatePolicyAssetAssociation = "create_policy_asset_association"
+const ReadPolicyAssetAssociation = "read_policy_asset_association"
+const DeletePolicyAssetAssociation = "delete_policy_asset_association"
+
+const CreateAccount = "create_account"
+const ReadAccount = "read_account"
+const UpdateAccount = "update_account"
+const DeleteAccount = "delete_account"
+
+const CreateSubAccount = "create_sub_account"
+const ReadSubAccount = "read_sub_account"
+const DeleteSubAccount = "delete_sub_account"
+
+const ReadAccountDataStorageRegion = "read_account_data_storage_region"
+const UpdateAccountDataStorageRegion = "update_account_data_storage_region"
+
+const ReadSiteMasking = "read_site_masking"
+const UpdateSiteMasking = "update_site_masking"
+
+const UpdateLogLevel = "update_log_level"
+
+const ReadSitePerformance = "read_site_performance"
+const UpdateSitePerformance = "update_site_performance"
+
+const CreateApiSecApiConfig = "create_api_sec_api_config"
+const ReadApiSecApiConfig = "read_api_sec_api_config"
+const UpdateApiSecApiConfig = "update_api_sec_api_config"
+const DeleteApiSecApiConfig = "delete_api_sec_api_config"
+
+const ReadApiSecEndpointConfig = "read_api_sec_endpoint_config"
+const UpdateApiSecEndpointConfig = "update_api_sec_endpoint_config"
+
+const ReadApiSecSiteConfig = "read_api_sec_site_config"
+const UpdateApiSecSiteConfig = "update_api_sec_site_config"
+
+const CreateCacheRule = "create_cache_rule"
+const ReadCacheRule = "read_cache_rule"
+const UpdateCacheRule = "update_cache_rule"
+const DeleteCacheRule = "delete_cache_rule"
+
+const CreateIncapRule = "create_incap_rule"
+const ReadIncapRule = "read_incap_rule"
+const UpdateIncapRule = "update_incap_rule"
+const DeleteIncapRule = "delete_incap_rule"
+
+const UpdateSecurityRule = "update_security_rule"
+
+const CreateSecurityRuleException = "create_security_rule_exception"
+const ReadSecurityRuleException = "read_security_rule_exception"
+const UpdateSecurityRuleException = "update_security_rule_exception"
+const DeleteSecurityRuleException = "delete_security_rule_exception"
+
+const CreateCustomCertificate = "create_custom_certificate"
+const ReadCustomCertificate = "read_custom_certificate"
+const UpdateCustomCertificate = "update_custom_certificate"
+const DeleteCustomCertificate = "delete_custom_certificate"
+
+const CreateDataCenter = "create_data_center"
+const ReadDataCenter = "read_data_center"
+const UpdateDataCenter = "update_data_center"
+const DeleteDataCenter = "delete_data_center"
+
+const CreateDataCenterConfiguration = "create_data_configuration"
+const ReadDataCenterConfiguration = "read_data_configuration"
+
+const ReadDataStorageRegion = "read_data_storage_region"
+const UpdateDataStorageRegion = "update_data_storage_region"
+
+const UpdateOriginPop = "update_origin_pop"
+
+const CreateDataCenterServer = "create_data_center_server"
+const UpdateDataCenterServer = "update_data_center_server"
+const DeleteDataCenterServer = "delete_data_center_server"
+
+const CreateTxtRecord = "create_txt_record"
+const ReadTxtRecord = "read_txt_record"
+const UpdateTxtRecord = "update_txt_record"
+const DeleteTxtRecord = "delete_txt_record"
+
+const ReadCspSiteConfiguration = "read_csp_site_configuration"
+const UpdateCspSiteConfiguration = "update_csp_site_configuration"
+
+const CreateCspSiteDomain = "create_csp_site_domain"
+const ReadCspSiteDomain = "read_csp_site_domain"
+const UpdateCspSiteDomain = "update_csp_site_domain"
+const DeleteCspSiteDomain = "delete_csp_site_domain"
+
+const CreateNotificationCenterPolicy = "create_notification_center_policy"
+const ReadNotificationCenterPolicy = "read_notification_center_policy"
+const UpdateNotificationCenterPolicy = "update_notification_center_policy"
+const DeleteNotificationCenterPolicy = "delete_notification_center_policy"